feat: Add string operators and negation
